+++ resources/ts3_client_install.rb
@@ -1,4 +1,0 @@
-resource_name :ts3_client_install
-
-property :install_dir, String, required: true
-property :version, String, required: true