+++ test/integration/server/server_spec.rb
@@ -8,3 +8,13 @@ describe port(9987) do
   it { should be_listening }
   its('protocols') { should eq ['udp'] }
 end
+
+describe port(10011) do
+  it { should be_listening }
+  its('protocols') { should eq ['tcp'] }
+end
+
+describe port(30033) do
+  it { should be_listening }
+  its('protocols') { should eq ['tcp'] }
+end