+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 vNext
 
 - Switched to 12.5-style for resource definition for the `ts3_install` and `ts3_configure` resources
+- Added testing and support for systemd compatible versions of Debian and CentOS
+- Added SysV init support in `ts3_server_configure`
 
 v 0.2.0 (16 June 2016)